[refac] item.repository 추가
- item.service 추상화함
